Refactor endpoints to routers
